fix(performance): Implement Suspense and ErrorBoundary with @suspensive/react
This commit resolves a Next.js performance warning (CSR bailout) and improves the application's resilience by implementing proper loading and error states using the @suspensive/react library.

### Key Changes:

-   **Dependency:** Added @suspensive/react to manage Suspense and ErrorBoundary declaratively.

-   **Suspense Boundary (products/page.tsx):**
    -   The client-side logic was extracted into a ProductsView component.
    -   This component is now wrapped with <Suspense> from suspensive, using GridSkeleton as a fallback. This prevents the page from client-side rendering while still allowing the use of useSearchParams.

-   **ErrorBoundary Setup:**
    -   ProductsView is also wrapped with <ErrorBoundary> from suspensive.
    -   This provides a granular error fallback for data fetching failures within the product list, without crashing the entire page.
    -   A global error.tsx was also added to act as a final catch-all for any unhandled errors.

This new structure not only fixes the build warning but also establishes a stable pattern for handling asynchronous states throughout the application.
