+++ src/app/products/[id]/page.tsx
@@ -1,6 +1,8 @@
 import { notFound } from 'next/navigation';
 import { ProductDetailCard } from '@/components/products';
 import { getProductById } from '@/lib/data';
+import { ErrorBoundary, Suspense } from '@suspensive/react';
+import { DetailPageSkeleton } from '@/components/common';
 
 interface ProductDetailPageProps {
   params: Promise<{ id: string }>;
@@ -15,8 +17,12 @@ export default async function ProductDetailPage({ params }: ProductDetailPagePro
   }
 
   return (
-    <div className='p-20 text-center'>
-      <ProductDetailCard product={product} />
-    </div>
+    <ErrorBoundary fallback={<div className='m-4 pt-14'>Error loading products</div>}>
+      <Suspense fallback={<DetailPageSkeleton />}>
+        <div className='p-20 text-center'>
+          <ProductDetailCard product={product} />
+        </div>
+      </Suspense>
+    </ErrorBoundary>
   );
 }
