+++ src/components/search/SearchModal.tsx
@@ -2,6 +2,7 @@
 
 import { closeSearchModal, useAppDispatch, useAppSelector } from '@/RTK';
 import { SearchInput } from '@/components/search';
+import { ErrorBoundary } from '@suspensive/react';
 
 export function SearchModal() {
   const dispatch = useAppDispatch();
@@ -17,7 +18,9 @@ export function SearchModal() {
         className='side-bar fixed top-0 left-0 w-96 p-4 pt-14 lg:w-xl'
         onClick={(e) => e.stopPropagation()}
       >
-        <SearchInput />
+        <ErrorBoundary fallback={<div className='m-4'>Unexpected error, please reload</div>}>
+          <SearchInput />
+        </ErrorBoundary>
       </div>
     </div>
   );
